Refactored package name, added application.yaml
