+++ pom.xml
@@ -12,7 +12,7 @@
 	<artifactId>jma-gateway</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>jma-gateway</name>
-	<description>courses</description>
+	<description>Zuul Gateway Proxy</description>
 	<properties>
 		<java.version>11</java.version>
 	</properties>
