+++ src/main/java/ru/jma/gateway/JmaGatewayApplication.java
@@ -1,4 +1,4 @@
-package ru.education_platform_jma.jmagateway;
+package ru.jma.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
