+++ src/main/resources/application.yaml
@@ -1,0 +1,3 @@
+spring:
+  application:
+    name: gateway
